The testing code, test_stdlib_logger.f90, for stdlib_logger.f90
The current testing code for stdlib_logger.f90. Its quite a bit of a mess as it tries to address all the options that don't result in an error stop, and the flexibility of the code result in quite a few disparate options. Note that running this code will generate six files, named first_log_file.txt, second_log_file.txt, .... Running it a second time will result in slightly different behavior as the tests include detecting not being able to replace an existing file. The files need to be examined to verify that their contents are as described in their contents. Addressing the failure modes that result in an error stop will require one file for each mode. Note that there are three procedures that have defined failure modes: add_log_file, add_log_unit, and remove_log_unit, and add_log_file and add_log_unit have several ways that they can fail so testing them requires about nine more files.